findobject_pr/requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-bootstrap5==24.2
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==1.26.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
thumbnail==1.5
